requirements.txt
azure-functions==1.20.0
azure-identity==1.17.1
azure-storage-blob==12.23.1
openai>=1.40.0
azure-cognitiveservices-speech==1.37.0
requests>=2.31.0
